Add auth and register users and midlleware auth for Routing

+++ resources/views/oneArticle.blade.php
@@ -2,6 +2,8 @@
 
 @section('content')
         <h3>{{ $article->title }}</h3>
-        <p><small><a href="/edit/{{$article->id}}">Изменить статью</a></small> <small><a href="/delete/{{$article->id}}">Удалить статью</a></small></p>
+        @if (!Auth::guest())
+                <p><small><a href="/edit/{{$article->id}}">Изменить статью</a></small> <small><a href="/delete/{{$article->id}}">Удалить статью</a></small></p>
+        @endif
         <p>{!! $article->content !!}</p>
 @endsection
